modules/location-api/requirements.txt
certifi==2021.10.8
DateTime==4.3
GeoAlchemy2==0.9.4
greenlet==1.1.2
grpcio==1.41.1
grpcio-tools==1.41.1
packaging==21.2
pep8==1.7.1
protobuf==3.19.1
pyparsing==2.4.7
pytz==2021.3
Shapely==1.8.0
six==1.16.0
SQLAlchemy==1.4.26
zope.interface==5.4.0
